service_auth/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-dotenv
drf-spectacular
psycopg2-binary